+++ src/app/inner/home/checkout/checkout.component.ts
@@ -41,12 +41,7 @@ export class CheckoutComponent  implements OnInit {
       this.logoutForm.patchValue({user_id:this.id})
       this.logoutForm.patchValue({logged_in_from:"mobile"})
       if(this.logoutForm.invalid){
-        //console.log("Invalid");	
-        localStorage.clear()
-        this.api.showSuccess('Logout successful')
-        // window.location.reload();                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
-        this.router.navigate(['../outer'])
-        this.close()
+        this.api.showSuccess('User Id Required')
       }
       else{
         this.logOut()
@@ -56,14 +51,16 @@ export class CheckoutComponent  implements OnInit {
         this.api.showLoading()
         this.api.logout(this.logoutForm.value).subscribe(
         (resp:any)=>{
+          if(resp){
+          
           localStorage.clear()
           this.logoutForm.reset()
           this.close()
-         
           this.api.showSuccess(resp.message)
-          // window.location.reload();
           this.router.navigate(['../outer'])
+          localStorage.clear()
           this.api.loaderDismiss()
+          }
         },
         (error:any)=>{
           this.api.loaderDismiss()
